separate modules in store
