+++ package.json
@@ -20,7 +20,8 @@
     "vue-material": "^1.0.0-beta-7",
     "vue-router": "^3.0.1",
     "vuelidate": "^0.6.1",
-    "vuex": "^3.0.1"
+    "vuex": "^3.0.1",
+    "vuex-router-sync": "^5.0.0"
   },
   "devDependencies": {
     "autoprefixer": "^7.1.2",
