+++ src/store/mutation-types.js
@@ -1,0 +1,10 @@
+// AUTHORIZATON & AUTHENTICATION
+export const AUTH_USER = 'AUTH_USER'
+export const CLEAR_AUTH_DATA = 'CLEAR_AUTH_DATA'
+export const CLEAR_AUTH_ERROR = 'CLEAR_AUTH_ERROR'
+export const AUTH_ERROR = 'AUTH_ERROR'
+export const STORE_USER = 'STORE_USER'
+
+// VESSELS
+export const STORE_VESSEL = 'STORE_VESSEL'
+export const DELETE_VESSEL = 'DELETE_VESSEL'
